requirements.txt
pandas==1.5.3
matplotlib==3.6.3
scikit-learn==1.2.1
catboost==1.1.1
lightgbm
mlflow-skinny==2.1.1
optuna==3.1.0
hydra-core==1.3.1
hydra-optuna-sweeper==1.2.0
